+++ src/phoenix/server/api/mutations/__init__.py
@@ -1,8 +1,8 @@
 import strawberry
 
 from phoenix.server.api.mutations.dataset_mutations import DatasetMutation
+from phoenix.server.api.mutations.export_events_mutations import ExportEventsMutation
 from phoenix.server.api.mutations.project_mutations import ProjectMutation
-from phoenix.server.api.types.ExportEventsMutation import ExportEventsMutation
 
 
 @strawberry.type
